Travis: Refactor out ds3_share_open_delete_ds2
Large methods need to be in separate modules.
